requirements.txt
asgiref==3.8.1
diff-match-patch==20241021
Django==4.0.1
django-debug-toolbar==3.2.4
django-environ==0.11.2
django-import-export==4.1.1
django-jazzmin==2.6.0
django-phonenumber-field==8.0.0
django-social-share==2.3.0
django-tinymce==4.1.0
et_xmlfile==2.0.0
openpyxl==3.1.5
phonenumbers==8.13.44
pillow==10.4.0
psycopg2-binary==2.9.9
sqlparse==0.5.1
tablib==3.5.0
typing_extensions==4.12.2
ua-parser==0.18.0
user-agents==2.2.0
whitenoise==6.7.0